requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
contractions==0.0.25
cycler==0.10.0
idna==2.10
importlib-metadata==2.0.0
inflect==4.1.0
joblib==0.17.0
kiwisolver==1.2.0
matplotlib==3.3.2
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.3
Pillow==7.1.2
pprintpp==0.4.0
pyahocorasick==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
textsearch==0.0.17
tqdm==4.51.0
Unidecode==1.1.1
urllib3==1.25.11
wincertstore==0.2
wordcloud==1.8.0
zipp==3.4.0
